Supports creation of a VMware appliances in OVA format (#4)
TL;DR
-----

Adds OVA builds

Details
-------

Updates the image creation to not only provide for AMIs but also OVAs.
OVA creationg enable vendors to ship a virtual appliace to vSphere
environments. Customers can install the virtual appliance and get a
running instance directly without having to run install. The image
contains the airgap install for the application and takes parameters for
the admin console password and license file. On first boot, a job runs
that completes the installation by moving the license file into place
and running the install. The job that runs the install is triggered if
the Embedded Cluster binary isn't located in it's final install
location. This allows for re-running the job after a `reset`.

This approach to the OVA depends on a few things that aren't present on the AMI:

* An `install.sh` script in the default user's home directory and handles the install.

* A SystemD unit to handle the one-time install, the unit invokes
  `install.sh`.

* Reading OVF parameters for the license file and the admin console
  password. These are configured in the OVA.

* There's a python script included in the build process to add the
  parameters to the OVF before creating the OVA.

Since the OVA build requires a few more dependencies than the AMI
approach, I added `shell.nix` and `Brewfile` to support installing
dependencies based on how you prefer to install them. If you're not
using Nix you can ignore `shell.nix`, and if you're not using Homebrew
you can ignore the `Brewfile`.